Move record and child-process helpers in with the agent
Both sat at the package root without belonging there.

_child_process is imported only by agent/tools/_exec.py and
agent/workspace/local.py. It goes to agent/ rather than either of those,
because those two are siblings and neither should import the other.

Record is an agent-run trace: attempts, cost, token_usage, events. The
compiler had one opportunistic write into it -- guarded by "if the file
happens to exist" -- which was the compiler reaching up into the agent's
file. texture_run already reports the usdz it exported, so the CLI that
drives it now points the record at the result. The compiler no longer
imports from agent at all.

The worker test that asserted on record.json becomes a test that the worker
reports the right usdz; the record update is covered at the CLI, where it
now lives.
