+++ tests/test_agent_harness.py
@@ -28,10 +28,10 @@ from mini_articraft.agent.harness import (
     _read_sdk_quickstart,
     _run_id_for_prompt,
 )
+from mini_articraft.agent.record import Record, read_conversation
 from mini_articraft.agent.tools import Tool, ToolContext, ToolResult
 from mini_articraft.agent.tools._core import workspace_digest
 from mini_articraft.agent.workspace.local import DEFAULT_MAIN_PY, LocalWorkspace
-from mini_articraft.record import Record, read_conversation
 
 
 class CompactingScriptedModel(ScriptedModel):
