+++ tests/test_compile.py
@@ -11,13 +11,13 @@ import pytest
 from PIL import Image
 
 import mini_articraft.agent.workspace.local as local_module
+from mini_articraft.agent.record import Record, read_conversation
 from mini_articraft.agent.workspace.local import (
     DEFAULT_MAIN_PY,
     LocalWorkspace,
     _run_isolated_process,
 )
 from mini_articraft.compiler.worker import _merge_test_reports, _serialize_test_report
-from mini_articraft.record import Record, read_conversation
 from mini_articraft.sdk import TestFailure, TestReport
